Встраивание базы данных в сборку

+++ Threats.Data/Entities/Negative/NegativeType.cs
@@ -2,7 +2,7 @@ namespace Threats.Data.Entities;
 
 public class NegativeType : Entity
 {
-    public NegativeType(long id, string name) : base((int)id)
+    public NegativeType(int id, string name) : base(id)
     {
         Name = name;
     }
